+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for protocol version 5. This adds the following features
   - support for smart contract upgradability
+  - a new representation of accounts that is better optimised for common operations
 - Fix an issue where the catch-up downloader would fail at a protocol update.
 
 ## 4.5.0